fix: Project carousel indicators failure
The little buttons under project photos in modal were not working right. This is
due to the conflict because of bootstrap classes. One of the indicators and
classes has to have a class active. But the ng-repeat of Angular JS is not fit
for it. I had to create a separate indicator and image div for the first project
screenshot.